+++ WaterTempTransmitter.h
@@ -15,17 +15,17 @@ public:
     // pin is the pin connected to the data of the 433 transmitter
     // id is an 8 bit identifier (you can create your own)
     // The channel which can be (1,2,3)
-    WaterTempTransmitter(uint8_t pin, uint8_t id, uint8_t channel);
+    WaterTempTransmitter(byte pin, byte id, byte channel);
 
-    void send(float temperature,bool battery=false, bool beep=false);
+    void send(float temperature, bool battery=false, bool beep=false);
 
 protected:
-    void sendPacket(uint8_t pin, uint32_t packet, uint8_t nbits);
-    void sendFrame(uint8_t pin, uint32_t packet, uint8_t nbits, uint8_t repetitions);
+    void sendPacket(byte pin, unsigned long packet, byte nbits);
+    void sendFrame(byte pin, unsigned long packet, byte nbits, byte repetitions);
 protected:
-    uint8_t _pin; 
-    uint8_t _id; 
-    uint8_t _chan;
+    byte _pin; 
+    byte _id; 
+    byte _chan;
 };
 
 #endif
